+++ CHANGELOG.md
@@ -13,6 +13,14 @@
 ### Changed
 
 - `all-is-cubes` library:
+    - Breaking: The `Block` type has been substantially redesigned (though it still has existing functionality).
+        - It is no longer an `enum`; it is an opaque type.
+        - It is logically made up of two components, a `block::Primitive` and any number of `block::Modifier`s.
+            - `block::Primitive` is an enum which has `Atom`, `Recur`, and `Indirect` variants as before.
+            - The `Rotated` variant is now `Modifier::Rotate`.
+        - It internally uses reference counting to be predictably cheap to `clone()`; this is intended to help use cases such as transactions which mean that even a block which only exists once in the universe is frequently cloned.
+        - Many changes to the functions, associated methods, and `BlockBuilder` were made to support this new structure.
+
     - `AllIsCubesAppState::new()` and `BlockProvider::new()` are now async functions.
     - The `linking::BlockModule` trait now requires the [`exhaust::Exhaust`](https://docs.rs/exhaust/latest/exhaust/trait.Exhaust.html) trait in place of `strum::IntoEnumIterator`. This allows implementors to use enums with fields (or non-enums).